Fix merge

+++ planetiler-core/src/main/java/com/onthegomap/planetiler/util/TilesetSummaryStatistics.java
@@ -232,9 +232,9 @@ public class TilesetSummaryStatistics {
 
     @Override
     public int compareTo(TileSummary o) {
-      int result = Integer.compare(archivedSize, o.archivedSize);
+      int result = Long.compare(archivedSize, o.archivedSize);
       if (result == 0) {
-        result = Integer.compare(coord.encoded(), o.coord.encoded());
+        result = Long.compare(coord.encoded(), o.coord.encoded());
       }
       return result;
     }
